完成涨跌通知功能，补充板块统计刷新
