+++ src/renderer/components/Home/FundList/EditFundContent/index.scss
@@ -11,6 +11,7 @@
     margin-bottom: 5px;
     display: block;
     font-weight: 600;
+    white-space: nowrap;
   }
 
   span {
