+++ src/renderer/components/Home/QuotationList/FundFlowContent/SouthFlow/index.tsx
@@ -1,8 +1,8 @@
-import React, { useState } from 'react';
+import React, { useCallback } from 'react';
 import { useRequest } from 'ahooks';
 
 import { useHomeContext } from '@/components/Home';
-import TypeSelection from '@/components/TypeSelection';
+import ChartCard from '@/components/Card/ChartCard';
 import { useResizeEchart, useRenderEcharts } from '@/utils/hooks';
 import * as CONST from '@/constants';
 import * as Services from '@/services';
@@ -106,10 +106,16 @@ const SouthFlow: React.FC<SouthFlowProps> = () => {
     [darkMode]
   );
 
+  const freshChart = useCallback(() => {
+    runGetFlowFromEastmoney(fields1, code);
+  }, [fields1, code]);
+
   return (
-    <div className={styles.content}>
-      <div ref={chartRef} style={{ width: '100%' }} />
-    </div>
+    <ChartCard onFresh={freshChart}>
+      <div className={styles.content}>
+        <div ref={chartRef} style={{ width: '100%' }} />
+      </div>
+    </ChartCard>
   );
 };
