+++ src/renderer/helpers/time.ts
@@ -2,10 +2,16 @@ import * as Adapter from '@/utils/adpters';
 import * as Services from '@/services';
 
 export async function GetCurrentHours() {
-  const collectors = [
-    Services.Time.GetCurrentDateTimeFromTaobao,
-    Services.Time.GetCurrentDateTimeFromJd,
-    Services.Time.GetCurrentDateTimeFromSuning,
-  ];
-  return Adapter.ChokePreemptiveAdapter<string>(collectors);
+  const now = Date.now().toString();
+  try {
+    const collectors = [
+      Services.Time.GetCurrentDateTimeFromTaobao,
+      Services.Time.GetCurrentDateTimeFromJd,
+      Services.Time.GetCurrentDateTimeFromSuning,
+    ];
+    return (await Adapter.ChokePreemptiveAdapter<string>(collectors)) || now;
+  } catch (error) {
+    console.log('获取远程时间出错', error);
+    return now;
+  }
 }
